third_party/blink/web_tests/wpt_internal/geolocation-api/permission-denied.https.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-adapter.js"></script>
</head>
<body>
<script type="module">
import {GeolocationMock} from './resources/geolocation-mock.js';

description("Tests Geolocation when permission is denied, using the mock service.");

let error;

const mock = new GeolocationMock();
mock.setGeolocationPermission(false);
mock.setGeolocationPosition(51.478, -0.166, 100.0);

navigator.geolocation.getCurrentPosition(function(p) {
    testFailed('Success callback invoked unexpectedly');
    finishJSTest();
}, function(e) {
    error = e;
    assert_equals(error.code, error.PERMISSION_DENIED);
    assert_equals(error.message, "User denied Geolocation");
    finishJSTest();
});
</script>
</body>
</html>
